Extend the game's state machine
We now use the state machine for everything,
including the game over state etc.

Every member of the GAME_STATES table must provide
an update() and a draw() function. If we set the
CURRENT_STATE variable to one of the known states,
we internally automatically choose the right functions
for the game loop.